# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rechvix API
  version: "0.1.0"
  description: >
    Multi-company, multi-branch, multi-warehouse billing, inventory,
    accounting, and India-GST tax-management API. This spec is generated
    by reading `internal/modules/*/httpapi/handlers.go` directly — every
    path, method, and request/response shape below is verified against
    the actual Go handler code, not inferred from `docs/architecture.md`'s
    design intent. See `docs/api.md` for narrative conventions (auth,
    error envelope, money shape, pagination status) that apply across
    every endpoint below rather than being repeated per-operation.

    Status: complete — all 124 real routes across all 15 HTTP modules are
    present and typed, mechanically verified against a grep of every
    r.Get/Post/Put/Delete/Patch(...) registration in the codebase (exact
    match, see the git history for this file). Kept current as endpoints
    ship (docs/TODO.md's cross-cutting checklist item) — re-run that same
    grep diff after adding or changing a route, don't just trust this
    note.
servers:
  - url: /api/v1
    description: Same origin as the web app; also reachable directly on the configured HTTP_PORT.
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /api-keys:
    post:
      tags: [Identity]
      summary: Create an API key. The raw key is returned exactly once — subsequent GETs return key_prefix only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/APIScope" }
                expires_at: { type: string, format: date-time, nullable: true }
                allowed_ip: { type: string, nullable: true, description: Single IP the key is restricted to, if set. }
      responses:
        "201":
          description: Created. `key` is the only response that ever contains the raw secret.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  key_prefix: { type: string }
                  scopes: { type: array, items: { $ref: "#/components/schemas/APIScope" } }
                  key: { type: string, description: Raw API key — store it now, it is never shown again. }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      tags: [Identity]
      summary: List API keys for the current user (key_prefix only, never the raw key).
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        key_prefix: { type: string }
                        scopes: { type: array, items: { $ref: "#/components/schemas/APIScope" } }
                        expires_at: { type: string, format: date-time, nullable: true }
                        last_used_at: { type: string, format: date-time, nullable: true }
                        created_at: { type: string, format: date-time }
  /api-keys/{id}:
    delete:
      tags: [Identity]
      summary: Revoke an API key immediately.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string, example: revoked } } }
        "404": { $ref: "#/components/responses/NotFound" }

  # ============================================================
  # ORGANISATION — tenant hierarchy: org / legal entity / branch / warehouse
  # Response bodies here are domain structs with NO json tags, so field
  # names serialize as Go's exact exported (PascalCase) names — unlike
  # every request body in this API, which is snake_case. Verified against
  # internal/modules/organisation/domain/domain.go.
  # ============================================================
...
  /contacts/parties:
    get:
      tags: [Contacts]
      summary: List, or search with `q`, parties.
      parameters:
        - { name: q, in: query, schema: { type: string }, description: "Free-text (name/phone); when present, results are capped at 20." }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { parties: { type: array, items: { $ref: "#/components/schemas/Party" } } } } } } }
    post:
      tags: [Contacts]
      summary: Create a party (customer, supplier, or both).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [party_type, legal_name]
              properties:
                party_type: { type: string, enum: [CUSTOMER, SUPPLIER, BOTH] }
                legal_name: { type: string }
                trade_name: { type: string }
                phone: { type: string }
                email: { type: string }
                currency_code: { type: string }
                credit_limit_amount: { type: string, nullable: true, description: Decimal string. }
                payment_terms_days: { type: integer, nullable: true }
                notes: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Party" } } } }
        "400":
          description: INVALID_PARTY_TYPE / LEGAL_NAME_REQUIRED / INVALID_CREDIT_LIMIT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /contacts/parties/{id}:
    get:
      tags: [Contacts]
      summary: Get a party by ID.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Party" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /contacts/parties/{id}/addresses:
    get:
      tags: [Contacts]
      summary: List a party's addresses.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { addresses: { type: array, items: { $ref: "#/components/schemas/Address" } } } } } } }
    post:
      tags: [Contacts]
      summary: Add an address to a party.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_type]
              properties:
                address_type: { type: string, enum: [BILLING, SHIPPING, WAREHOUSE, REGISTERED_OFFICE] }
                line1: { type: string }
                line2: { type: string }
                city: { type: string }
                state: { type: string }
                postal_code: { type: string }
                country_code: { type: string }
                is_default: { type: boolean }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Address" } } } }
        "400":
          description: INVALID_ADDRESS_TYPE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /contacts/parties/{id}/tax-registrations:
    get:
      tags: [Contacts]
      summary: List a party's tax registrations (e.g. GSTIN).
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { tax_registrations: { type: array, items: { $ref: "#/components/schemas/TaxRegistration" } } } } } } }
    post:
      tags: [Contacts]
      summary: Add a tax registration to a party.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [country_code, registration_number]
              properties:
                country_code: { type: string }
                registration_number: { type: string }
                state_code: { type: string }
                is_primary: { type: boolean }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/TaxRegistration" } } } }
        "409":
          description: DUPLICATE_REGISTRATION.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /contacts/tax-registrations/{number}:
    get:
      tags: [Contacts]
      summary: Look up a party by its tax registration number (e.g. GSTIN reverse lookup).
      parameters: [{ name: number, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/TaxRegistration" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /contacts/parties/import:
    post:
      tags: [Contacts]
      summary: Bulk-import customers/suppliers from an uploaded CSV or XLSX file.
      parameters:
        - { name: format, in: query, required: true, schema: { type: string, enum: [csv, xlsx] } }
        - { name: dry_run, in: query, schema: { type: boolean, default: false } }
      requestBody:
        required: true
        content:
          text/csv: { schema: { type: string, format: binary } }
          application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: { schema: { type: string, format: binary } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ImportReport" } } } }
        "400":
          description: INVALID_CSV / INVALID_XLSX / INVALID_FORMAT.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  # ============================================================
  # PRICING — price lists and per-variant/unit price items
  # Verified against internal/modules/pricing/domain/domain.go.
  # ============================================================
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bp_session
      description: Set by POST /auth/login. Default cookie name; configurable via SESSION_COOKIE_NAME.
    apiKey:
      type: http
      scheme: bearer
      description: >
        Authorization: Bearer <key>. Effective permissions are the
        intersection of the key's declared scopes and the owning user's
        real RBAC grants — see docs/api.md.
  responses:
...
    BadRequest:
      description: Malformed request body or invalid field value.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: Resource does not exist, or exists in a different organisation (never distinguished from the client's perspective).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: SESSION_INVALID }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    APIScope:
      type: string
      enum: [products:read, inventory:read, customers:read, customers:write, invoices:read, invoices:write, reports:read]
...
    ImportReport:
      type: object
      description: Shared by every bulk-import endpoint (products, parties, ...) — internal/platform/importer.Report.
      properties:
        DryRun: { type: boolean }
        Total: { type: integer }
        Committed: { type: integer }
        Valid: { type: integer, description: Valid but not committed (dry_run=true only). }
        Errors: { type: integer }
        Duplicates: { type: integer }
        Results:
          type: array
          items:
            type: object
            properties:
              RowNumber: { type: integer, description: 1-based, counting from the first data row (header excluded). }
              Outcome: { type: string, enum: [COMMITTED, VALID, ERROR, DUPLICATE] }
              Message: { type: string, description: Populated for ERROR (why) and DUPLICATE (what it matched). }
    Party:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        PartyType: { type: string, enum: [CUSTOMER, SUPPLIER, BOTH] }
        LegalName: { type: string }
        TradeName: { type: string }
        Phone: { type: string }
        Email: { type: string }
        CurrencyCode: { type: string }
        CreditLimitAmount: { type: string, nullable: true }
        PaymentTermsDays: { type: integer, nullable: true }
        Notes: { type: string }
        Status: { type: string }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    Address:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        PartyID: { type: string, format: uuid }
        AddressType: { type: string, enum: [BILLING, SHIPPING, WAREHOUSE, REGISTERED_OFFICE] }
        Line1: { type: string }
        Line2: { type: string }
        City: { type: string }
        State: { type: string }
        PostalCode: { type: string }
        CountryCode: { type: string }
        IsDefault: { type: boolean }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    TaxRegistration:
      type: object
      properties:
        ID: { type: string, format: uuid }
        OrganisationID: { type: string, format: uuid }
        PartyID: { type: string, format: uuid }
        CountryCode: { type: string }
        RegistrationNumber: { type: string }
        StateCode: { type: string }
        IsPrimary: { type: boolean }
        CreatedAt: { type: string, format: date-time }
...